requirements.txt
aiohttp
aiohttp_jinja2
attrs
jinja2
setuptools
pymongo[srv]
